+++ Source/Utility/Mat4f.h
@@ -158,12 +158,12 @@ namespace TrenchBroom {
                 return *this;
             }
             
-            inline float& operator[] (const unsigned int index) {
+            inline float& operator[] (const size_t index) {
                 assert(index >= 0 && index < 16);
                 return v[index];
             }
             
-            inline const float& operator[] (const unsigned int index) const {
+            inline const float& operator[] (const size_t index) const {
                 assert(index >= 0 && index < 16);
                 return v[index];
             }
