+++ Source/Utility/Vec2f.h
@@ -99,13 +99,13 @@ namespace TrenchBroom {
                 return *this;
             }
             
-            inline float& operator[] (const unsigned int index) {
+            inline float& operator[] (const size_t index) {
                 assert(index >= 0 && index < 2);
                 if (index == 0) return x;
                 return y;
             }
             
-            inline const float& operator[] (const unsigned int index) const {
+            inline const float& operator[] (const size_t index) const {
                 assert(index >= 0 && index < 2);
                 if (index == 0) return x;
                 return y;
